dating/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-filter==23.2
django-ipware==5.0.0
djangorestframework==3.14.0
geographiclib==2.0
idna==3.4
Markdown==3.4.3
Pillow==10.0.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
